+++ meshspy-ui/src/lib/MapContext.jsx
@@ -1,23 +1,63 @@
-// meshspy-ui/src/lib/MapContext.jsx
-import { createContext, useContext, useRef, useState } from "react";
+import React, { createContext, useContext, useRef, useState, useEffect } from "react";
+import { useNodes } from "../lib/api";
 
-const MapContext = createContext(null);
+const MapContext = createContext();
 
 export function MapProvider({ children }) {
   const mapRef = useRef(null);
   const markersRef = useRef({});
   const [isReady, setIsReady] = useState(false);
-  console.log("MapProvider isReady:", isReady);
+  const [selectedNodeId, setSelectedNodeId] = useState(null);
+  const [nodes, setNodes] = useState([]);
+
+  const { data: rawData } = useNodes();
+
+  useEffect(() => {
+    if (!rawData) return;
+
+    const result = Object.entries(rawData).map(([id, info]) => {
+      const payload = info.data?.payload ?? {};
+      const latRaw = payload.latitude_i ?? info.data?.latitude;
+      const lonRaw = payload.longitude_i ?? info.data?.longitude;
+
+      const lat = latRaw != null ? latRaw / 1e7 : null;
+      const lon = lonRaw != null ? lonRaw / 1e7 : null;
+
+      return {
+        id,
+        name:
+          payload.longname ||
+          payload.shortname ||
+          info.name ||
+          info.data?.name ||
+          id,
+        latitude: lat,
+        longitude: lon,
+        hasPosition: lat !== null && lon !== null,
+        raw: info,
+      };
+    });
+
+    setNodes(result);
+  }, [rawData]);
 
   return (
-    <MapContext.Provider value={{ mapRef, markersRef, isReady, setIsReady }}>
+    <MapContext.Provider
+      value={{
+        mapRef,
+        markersRef,
+        isReady,
+        setIsReady,
+        selectedNodeId,
+        setSelectedNodeId,
+        nodes,
+      }}
+    >
       {children}
     </MapContext.Provider>
   );
 }
 
-export function useMapContext() {
-  const ctx = useContext(MapContext);
-  if (!ctx) throw new Error("useMapContext must be used inside MapProvider");
-  return ctx;
+export function useMap() {
+  return useContext(MapContext);
 }
